refactor: POM and feature files
