+++ test/features/list_management.feature
@@ -10,8 +10,10 @@ Feature: List Management
     When the user clicks the continue button
     Then the user should be redirected to "https://trello.com/b/1s2GyOuN/mi-tablero-de-trello"
     When the user clicks on the create button
-    When selects create board
-    Then enters the board name "board" and clicks on the create button
+    When the user clicks on Create Board
+    And enters the board name "board" 
+    And clicks on the create button
     When the user clicks on Add a list
-    When enters the list name "list"  and clicks on the Add list button
-    Then the new list list_name should appear on the board
+    And enters the list name "list"  
+    And clicks on the Add list button
+    Then the new list should appear on the board
