+++ test/pages/CardMangementPage.js
@@ -1,22 +1,8 @@
+const cardManagementComponent = require('../components/CardManagementComponent');
+
 class CardManagmentPage {
-    get inputCardName() { return $('[data-testid="list-card-composer-textarea"]'); }
-    get addCardToListButton() { return $('[data-testid="list-card-composer-add-card-button"]'); }
-    cardElement(cardTitle) { return $(`a[data-testid="card-name"]=${cardTitle}`); }
-
-    async enterCardName(cardName) {
-        await this.inputCardName.waitForDisplayed();
-        await this.inputCardName.click();
-        await this.inputCardName.setValue(cardName);
-    }
-
-    async clickaddCardToList() {
-        await this.addCardToListButton.waitForDisplayed();
-        await this.addCardToListButton.click();
-    }
-
-    async isCardDisplayed(cardTitle) {
-        const card = this.cardElement(cardTitle);
-        return await card.isExisting() && card.isDisplayed();
+    constructor() {
+        this.cardManagement = cardManagementComponent;
     }
 }
 module.exports = new CardManagmentPage();
