+++ test/pages/ManageProfilePage.js
@@ -1,20 +1,9 @@
+const manageProfileComponent = require('../components/ManageProfileComponent');
+
 class ManageProfilePage {
     
-    get permissionsOptions() { return $('div[data-test-selector=visibility-wrapper-localtime]'); }
-    get permissionsAnyone() { return $('div=Anyone'); }
-
-
-    async editLocation() {
-        await this.permissionsOptions.waitForDisplayed();
-        await this.permissionsOptions.click();
-        await this.permissionsAnyone.waitForDisplayed();
-        await this.permissionsAnyone.click();
-    }
-
-    async switchToWindow() {
-        await browser.pause(2000);
-        const handles = await browser.getWindowHandles();
-        await browser.switchToWindow(handles[1]);
+    constructor() {
+        this.manageProfile = manageProfileComponent;
     }
 
 }
