+++ test/step-definitions/boardManagementSteps.js
@@ -7,30 +7,35 @@ const testData = require('../utils/testData');
 let createdBoardTitle;
 
 When(/^the user clicks on the create button$/, async () => {
-    await HomePage.ClickCreateMenu();
+    await HomePage.home.clickCreateMenu();
 });
 
-When(/^selects create board$/, async () => {
-    await HomePage.ClickCreateBoard();
+When(/^the user clicks on Create Board$/, async () => {
+    await HomePage.home.clickCreateBoard();
 });
 
-Then(/^enters the board name "([^"]*)" and clicks on the create button, the board is displayed into the board list$/, async (boardKey) => {
+Then(/^enters the board name "([^"]*)"$/, async (boardKey) => {
     createdBoardTitle = testData[boardKey].title;
-    await BoardManagementPage.enterBoardTitle(createdBoardTitle);
-    await BoardManagementPage.clickCreateBoard();
-    await BoardManagementPage.clickBoardListShow();
-    const boardElement = await BoardManagementPage.boardItem(createdBoardTitle);
+    await BoardManagementPage.board.enterBoardTitle(createdBoardTitle);
+});
+
+When(/^clicks on the create button$/, async () => {
+    await BoardManagementPage.board.clickCreateBoard();
+});
+
+Then(/^the board should be displayed into the board list$/, async () => {
+    await BoardManagementPage.board.clickBoardListShow();
+    const boardElement = await BoardManagementPage.board.boardItem(createdBoardTitle);
     await boardElement.waitForDisplayed({ timeout: 5000 });
     expect(await boardElement.isDisplayed()).to.be.true; 
 });
 
-Then(/^enters the board name "([^"]*)" and clicks on the create button, the board is displayed into the search list$/, async (boardKey) => {
-    createdBoardTitle = testData[boardKey].title;
-    await BoardManagementPage.enterBoardTitle(createdBoardTitle);
-    await BoardManagementPage.clickCreateBoard();
+Then(/^the board should be displayed into the search list$/, async () => {
     await browser.pause(3000);
-    await HomePage.clickBoardslink();
-    await BoardManagementPage.enterSearchBoard(createdBoardTitle);
-    const isBoardDisplayed = await BoardManagementPage.isBoardDisplayed(createdBoardTitle);
+    await HomePage.home.clickBoardslink();
+    await BoardManagementPage.board.enterSearchBoard(createdBoardTitle);
+    const isBoardDisplayed = await BoardManagementPage.board.isBoardDisplayed(createdBoardTitle);
     expect(isBoardDisplayed).to.be.true;
 });
+
+
