+++ test/step-definitions/cardManagementSteps.js
@@ -8,17 +8,17 @@ let createdCardTitle;
 
 When(/^the user clicks on Add a card under the list_name list$/, async () => {
     await browser.pause(3000);
-    await BoardManagementPage.clickAddCard();
+    await BoardManagementPage.board.clickAddCard();
 });
 
 When(/enters the card name "([^"]*)" and clicks on the add button$/, async (cardKey) => {
     createdCardTitle = testData[cardKey].title;
-    await CardManagmentPage.enterCardName(createdCardTitle);
-    await CardManagmentPage.clickaddCardToList(); 
+    await CardManagmentPage.cardManagement.enterCardName(createdCardTitle);
+    await CardManagmentPage.cardManagement.clickAddCardToList();
 });
 
 Then(/the new card card_name should appear on the board$/, async () => {
-    const card = await CardManagmentPage.cardElement(createdCardTitle);
+    const card = await CardManagmentPage.cardManagement.cardElement(createdCardTitle);
     await card.waitForExist(); 
     await card.waitForDisplayed();   
     expect(await card.isDisplayed()).to.be.true;
